+++ lib/http_headers/utils.rb
@@ -1,5 +1,7 @@
 require 'http_headers/utils/version'
+
 require 'http_headers/utils/list'
+require 'http_headers/utils/single'
 
 module HttpHeaders
   module Utils