+++ test/http_headers/utils/list_test.rb
@@ -1,18 +1,21 @@
-require "test_helper"
+require 'test_helper'
+require 'delegate'
 
 module HttpHeaders
   module Utils
     class ListTest < Minitest::Test
-
-      class TestAccept < HttpHeaders::Utils::List
+      class TestAccept < DelegateClass(Array)
 
         def initialize(value)
-          super value, entry_klazz: Entry
-          sort!
+          list = HttpHeaders::Utils::List.new(value, entry_klazz: Entry)
+          __setobj__ list
+        end
+
+        def to_s
+          HttpHeaders::Utils::List.to_header(self)
         end
 
         class Entry
-
           attr_reader :media_type, :index
 
           def initialize(media_type, index:, parameters:)
@@ -31,8 +34,12 @@ module HttpHeaders
             index <=> other.index
           end
 
+          def to_header
+            [media_type].concat(parameters.map { |k, v| "#{k}=#{v}" }).compact.reject(&:empty?).join('; ')
+          end
+
           def to_s
-            media_type
+            to_header
           end
 
           def [](index)
@@ -68,8 +75,13 @@ module HttpHeaders
         list = TestAccept.new(['*/*; q=0.1', 'application/json, text/html; q=0.8'])
         assert_equal 3, list.length
         assert_equal 'application/json', list.first.to_s
-        assert_equal '*/*', list.last.to_s
+        assert_equal '*/*; q=0.1', list.last.to_s
         assert_equal 0.8, list[1].q
+      end
+
+      def test_it_can_convert_to_a_header
+        list = TestAccept.new(['*/*; q=0.1', 'application/json, text/html; q=0.8'])
+        assert_equal 'application/json, text/html; q=0.8, */*; q=0.1', list.to_s
       end
     end
   end