+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Changelog
 
+## 0.2.0
+
+- Fix whitespace on input by using `strip`
+- Change `List` from `class` to `module`
+- Change `List.new` to automatically sort
+- Add `Single` for a single header value
+- Add `to_header` lookup and implementation for `Single` and `List`
+- Change `to_header` to use `HEADER_DELIMITER`
+
 ## 0.1.1
 
-Add a missing require to `delegate`
+- Add a missing require to `delegate`
 
 ## 0.1.0